refactor(Real-ESRGAN): Optimize cpp postprocess and some other improvements.
(cherry picked from commit b69ae704e354bb716ba4fdcbd2bd984da56925e5)

+++ sample/Real-ESRGAN/python/real_esrgan_opencv.py
@@ -146,7 +146,7 @@ class Real_ESRGAN:
         #cpu postprocess
         outputs_arr = outputs_arr.astype(np.float32)
         outputs_arr *= 255
-        outputs_arr = outputs_arr.transpose(0, 2, 3, 1)
+        outputs_arr = outputs_arr.astype(np.uint8).transpose(0, 2, 3, 1)
         clipped_arr = []
         for i in range(len(outputs_arr)):
             if txy_list[i][0] != 0: #tx
